backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
pydantic==2.4.2
httpx==0.25.0
python-dotenv==1.0.0
prometheus-client==0.17.1
pydantic-settings==2.0.3
requests==2.31.0